ARM: dts: set the compressed ramfs size to 24MB
In original implementation, commandline and filesystem size is specified
in ATAGS, not DTS. The serial console is specified with ttyAMA0, not
ttyS0. It's a bug.

Now we try to specify commandline, ramfs in DTS file. So we have to
specify the size of compressed ramfs in DTS file, too. It's inconvenient
when we update the ramfs. Now specify the compressed ramfs with 24MB.
Then we can avoid to always update .boot since the size of filesystem
is in .boot image. The size of ramfs is better not to exceed 24MB.
Otherwise, we always meet failure of downloading & burning in UEFI.

The command of making ramfs is in below:
find . | cpio --quiet -H newc -o | gzip -9 -n >
../../boot-wrapper/filesystem.cpio.gz

The command of padding ramfs to 24MB is in below:
dd conv=sync if=filesystem.cpio.gz of=24m.img bs=1M count=24

The command of renaming ramfs image is in below:
cp 24m.img .filesystem

.filesystem could be downloaded and burned into NAND flash by UEFI.

Signed-off-by: Haojian Zhuang <haojian.zhuang@gmail.com>